backend/requirements.txt
fastapi==0.111.1
uvicorn==0.39
requests==2.31.0
python-dotenv>=1.0.0
pyspark==3.5.1
redis